Set up static reference to handler functions in Rectangle constructor
This helps us to correctly remove the event listeners from the canvas
when they are not needed.

We also added a custom event to trigger when a rectangle becomes active,
this then needs to be handled at a higher level, i.e. within the SHAPE
module. This module keeps a track of all the rectangles on the canvas,
so it can set all other rectangles as inactive, based on this event.